[Update] dashboardController/ PostModel dashboard controller - 좋아요 데이터 같이 가져오기 - 게시글 수정 요청 : multer 함수 route로 이동 -  좋아요 데이터 요청할때 user_id도 사용 - 게시글 상세조회 / 게시글 수정 상세조회 분리 (수정 데이터 조회시 제목,본문만 가져옴) - 게시글 수정 시 기존 이미지 삭제 가능(isImageDeleted로 확인)
PostModel
- 게시글 목록 조회 시 계정 프로필 이미지도 같이 select
- 반환 result 값 수정
- 자잘한 버그 오류 수정

dashboardRoutes
- 게시글 수정 데이터 조회 api 추가

ImageHandler
- 기존 이미지 삭제 기능 추가 후 버그 수정

- Refactored dashboardController to improve post data retrieval, including like status and user profile images.
- Added new methods for fetching edit data and handling post additions with image uploads.
- Updated SQL queries in PostModel to include user profile images and improved error handling.
- Enhanced deletePost functionality to ensure image deletion and proper response handling.
- Cleaned up code for better readability and maintainability.

+++ server/Service/ImageHandler.js
@@ -19,6 +19,7 @@ export const deleteProfileImage = async (profile_image) => {
 
 //게시글 이미지 삭제
 export const deletePostImage = async (post_image) => {
+    if(!post_image) return;
     const imagePath = path.join(rootDir,'data/images/',post_image);
     try{
         await fs.promises.unlink(imagePath);
